xauth: update to 1.1.3.
This is a minor bugfix release, including fixes for compiler warnings.

For those building for 32-bit platforms, it also enables use of the
"large file" APIs - while Xauthority files should never be more than 2gb
in size, they may be stored on filesystems with large inodes.

Since this release was generated using the new GNU autoconf 2.72, this
also adds a --enable-year2038 configure flag which may allow xauth
to work with files whose timestamps are later than January 19, 2038, but
this has not been tested.

Alan Coopersmith (9):
      config: Add missing AC_CONFIG_SRCDIR
      Variable scope reductions, as suggested by cppcheck
      Remove unnecessary casts from malloc() calls
      Handle -Wempty-body warning for WRITES() macro
      configure: Use AC_SYS_LARGEFILE to enable large file support
      configure: raise minimum autoconf requirement to 2.70
      unifdef hpux
      unifdef Lynx
      xauth 1.1.3

наб (1):
      Unbold brackets for optional n at start of commands