gotosocial: update to version 0.14.2
Changelog:


v0.14.2 Spoiledest Sloth

Another day another bugfix release 😎

We fixed a bug in 0.14.0 and 0.14.1 that was causing 404 errors sometimes when viewing an account or its statuses, if that account had moved at some point.

No db migrations or config file changes between v0.14.1 and this version, so upgrading is as simple as updating your docker container or stopping GtS, untarring the release, and starting GtS again.

For instructions to update to this release from versions < v0.14.0, see the v0.14.0 release notes, but replace v0.14.0 with v0.14.2 throughout.

Thanks!
Detailed Changelog

    0bd95d7 [bugfix] Don't error when populating MovedTo if account not found (#2741)
    0326318 [chore] seperate snapshot from release in goreleaser (#2740)


v0.14.1 Spoileder Sloth

Hello! Here's a bugfix release!

There was a bug in 0.14.0 which was causing SQLite-backed instances to sometimes become unresponsive, with lots of errors in the logs containing interrupted(9) (see #2728).

We poked around and this looks like an issue in our SQLite dependency, so we stepped down to the most recent stable version, and also added some code to prevent the problem from rendering an instance unresponsive.

No db migrations or config file changes between v0.14.0 and this version, so upgrading is as simple as updating your docker container or stopping GtS, untarring the release, and starting GtS again.

For instructions to update to this release from versions < v0.14.0, see the v0.14.0 release notes, but replace v0.14.0 with v0.14.1 throughout.

Thanks!
Detailed Changelog

    ce3ae09 [chore] Downgrade sqlite v1.29.2 -> v1.28.0 (#2736)
    6418785 [bugfix] add workaround for Xsqlite_interrupt() permanently breaking connection (#2731)