libnbcompat: Rework in-place linking.
Prevent libtool from detecting -lnbcompat in the command line arguments
then declaring it as a dependency in the resulting libtool archive.

Filter -lnbcompat from all of the environment when compiling an
in-place copy to prevent circular dependencies.

Drop LIBNBCOMPAT_USE_PIC, I can't find anywhere this is used in the
tree, it's incomplete (needs handling in mk/) and complicates things
needlessly.

What still needs to be done is figure out why certain packages don't
seem to properly link with libnbcompat when USE_FEATURES is used - it's
a long-term problem, might be related to some newer build systems not
using ${LIBS}.

Bootstrap tested on Darwin and NetBSD.