qt6: updated to 6.6.3
Qt 6.6.3 release is a patch release made on the top of Qt 6.6.2.
As a patch release, Qt 6.6.3 does not add any new functionality but provides
bug fixes and other improvements and maintains both forward and backward
compatibility (source and binary) with Qt 6.6.2.

For detailed information about Qt 6.6, refer to the online documentation
included in this distribution. The documentation is also available online:

https://doc.qt.io/qt-6/index.html