xload: update to 1.2.0.
This release adds -help and -version options, and removes support for a
number of older OS'es that predate the common getloadavg() system call.

For those building for 32-bit platforms, it also enables use of the
"large file" APIs in case its necessary to access file whose size or
inode number is outside the range of the original 32-bit API.

Since this release was generated using the new GNU autoconf 2.72, this also
adds a --enable-year2038 configure flag which may allow xload on 32-bit
platforms to work with files whose timestamps are later than January 19, 2038,
but this has not been tested.

Alan Coopersmith (17):
      unifdef -UUSG
      gitlab CI: stop requiring Signed-off-by in commits
      Print which arguments were unknown before giving usage message
      Add -help and -version options
      Use asprintf() if available
      configure: Use AC_SYS_LARGEFILE to enable large file support
      unifdef SYSV
      unifdef __osf__
      unifdef hpux
      unifdef SVR4
      unifdef att
      unifdef QNX4
      Use autoconf to check for <paths.h> instead of ifdef BSD
      Use autoconf to check for <protocols/rwhod.h> instead of OS-specific ifdefs
      unifdef __bsdi__
      Use imdent to realign remaining pre-processer directives
      xload 1.2.0