ltm: update to 1.3.0.
This release is done in preparation for libtommath 2.0.0 which will
deprecate multiple public parts of the API, both functions and
macros.

Additionally we’ve included CMake as build system and use GitHub
Actions to verify that everything works as expected.

All to-be-deprecated parts are marked as either MP_DEPRECATED()
for functions or MP_DEPRECATED_PRAGMA() for macros. Your compiler
should complain if you use those to-be-deprecated parts, if not
it’s most likely a pretty old version of the compiler or not
correctly configured.