snac: update to 2.51. Changes:
Support for custom Emojis has been added; they are no longer hardcoded,
but read from the `emojis.json` file at the server base directory. Also,
they are no longer limited to string substitutions, but images as
external URLs are also supported (see `snac(8)` for more information).

Fixed a bug that caused some notifications to be lost when coming from a
user in the same instance.

Added an additional check for blocked instances (sometimes, posts from
blocked sites that were ancestors of legit posts were 'leaking' into the
timeline).

On OpenBSD, if the `disable_email_notifications` server flag is set to
`true`, `unveil()` is not called for the execution of the
`/usr/sbin/sendmail` binary and `pledge()` doesn't set the `exec`
promise.