py-scikit-build-core: updated to 0.9.0
Version 0.9.0

This version adds the ability to `inherit` in override tables, matching a
similar feature added to cibuildwheel 2.17's overrides. You can now write out
extra metadata to `@{SKBUILD_METADATA_DIR}`. A new Hatchling plugin is provided
as an experimental feature (will likely be made a separate package in the future
like the setuptools plugin).

The meaning of `wheel.exclude` has been changed to match on the wheel path,
rather than the source path. We could not find any projects that would be
affected by this change, so it was not added to the minimum-version protection
policy. This now allows you to ignore CMake outputs as well.

Features:

- Preserve additivity of `cmake.define` across `overrides` tables
- Add metadata dir access
- Experimental hatchling builder
- Vendor pyproject-metadata
- Always require pathspec

Fixes:

- Exclude installed files if listed in exclude
- Make `.git_archival.txt` reproducible
- Use `cmake -E` capabilities instead of `cmake --version`
- Ensure many/musl tags not selected
- purelib should set py3 tag if unset
- Validate description for 0.9+
- Support bools in config settings

API changes:

- `extra_settings` for SettingsReader
- `GetRequires` args changed
- Make `from_file` a little more powerful
- Metadata is part of the build backend

Documentation:

- `cmakelists.md` Windows `SOABI` suffix variable
- Fix hatch init command
- Fix install strip default
- Improve `ninja.make-fallback` description in the README
- Mention printouts

CI and testing:

- Lower `pybind11` test dependency
- Some cleanup from uv branch
- Fedora CI maintenance
- Small additions
- Some changes from uv job
- Fix setuptools on Python 3.12
- Fedora: Port downstream PR-49